requirements.txt
Django>=2.2.0
djangorestframework>=3.9.4

dateutils>=0.6.6
watchdog>=0.8.3
psutil>=5.0.1
tzlocal>=1.4
mutagen>=1.37
pyyaml>=3.12

django-filer>=1.5.0
django-admin-sortable2>=0.7.2
django-content-editor>=1.4.2
feincms3[all]>=0.31.0

bleach>=1.4.3
django-honeypot>=0.5.0

gunicorn>=19.6.0

